Iterate on data, not on axis data
It was already strange to iterate on one axis data.
On d3-select v2, it's not really possible anymore.
Because for instance, with async functions like “on”
("onmouseover" or others), d3 doesn't pass anymore
the index of the element we are on. Only the event
and current data. That makes impossible to retrieve
original data (as we iterate on one axis data).
This commit remove the axis data for a data object
with all axis data inside. That’s also simplify the
logic.